#143 fetching by ID

+++ JS/node/nodejs-complete-giude/controllers/shop.js
@@ -15,13 +15,21 @@ exports.getProducts = (req, res, next) => {
 
 exports.getProduct = (req, res, next) => {
   const prodId = req.params.productId;
-  Product.findById(prodId, product => {
-    res.render("shop/product-detail", {
-      product: product,
-      pageTitle: product.title,
-      path: "/products"
-    });
-  });
+
+  Product.findById(prodId)
+    .then(([rows]) => (rows.length > 0 ? rows[0] : null))
+    .then(product => {
+      if (product) {
+        res.render("shop/product-detail", {
+          product: product,
+          pageTitle: product.title,
+          path: "/products"
+        });
+      } else {
+        return res.redirect("/");
+      }
+    })
+    .catch(console.log);
 };
 
 exports.getIndex = (req, res, next) => {
@@ -65,18 +73,33 @@ exports.getCart = (req, res, next) => {
 
 exports.postCart = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.findById(prodId, product => {
-    Cart.addProduct(prodId, product.price);
-  });
-  res.redirect("/cart");
+
+  Product.findById(prodId)
+    .then(([rows]) => (rows.length > 0 ? rows[0] : null))
+    .then(product => {
+      if (product) {
+        Cart.addProduct(prodId, product.price);
+        res.redirect("/cart");
+      } else {
+        return res.redirect("/cart");
+      }
+    })
+    .catch(console.log);
 };
 
 exports.postCartDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.findById(prodId, product => {
-    Cart.deleteProduct(prodId, product.price);
-    res.redirect("/cart");
-  });
+  Product.findById(prodId)
+    .then(([rows]) => (rows.length > 0 ? rows[0] : null))
+    .then(product => {
+      if (product) {
+        Cart.deleteProduct(prodId, product.price);
+        res.redirect("/cart");
+      } else {
+        return res.redirect("/cart");
+      }
+    })
+    .catch(console.log);
 };
 
 exports.getOrders = (req, res, next) => {
